Move filesystem functions somewhere sane
